add getproduct by keyword
