+++ Backend/1015bookstore.application/Catalog/Products/IProductService.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace _1015bookstore.application.Catalog.Products
 {
-    public interface IManageProductService
+    public interface IProductService
     {
         Task<int> Create(ProductCreateRequest request);
         Task<int> Update(ProductUpdateRequest request);
@@ -17,6 +17,8 @@ namespace _1015bookstore.application.Catalog.Products
         Task<bool> UpdatePrice(int id, decimal newPrice);
         Task<bool> UpdataQuanity(int id, int addedQuantity);
         Task AddViewcount(int id);
-        Task<PagedResult<ProductViewModel>> GetAllPaging(GetManageProductPagingRequest request);
+        Task<PagedResult<ProductViewModel>> GetProductByKeyWordPaging(GetProductByKeyWordPagingRequest request);
+        Task<List<ProductViewModel>> GetAll();
+        Task<PagedResult<ProductViewModel>> GetProductByCategoryId(GetProductByCategoryPagingRequest request);
     }
 }
