--- Backend/1015bookstore.application/Catalog/Products/IPublicProductService.cs
@@ -1,16 +1,0 @@
-﻿using _1015bookstore.viewmodel.Catalog.Products;
-using _1015bookstore.viewmodel.Comon;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace _1015bookstore.application.Catalog.Products
-{
-    public interface IPublicProductService
-    {
-        Task<List<ProductViewModel>> GetAll();
-        Task<PagedResult<ProductViewModel>> GetAllByCategoryId(GetPublicProductPagingRequest request);
-    }
-}
